Added event coder recruiter Seeder and EventFactory CoderFactory RecruiterFactory

+++ database/migrations/2023_03_27_110346_create_events_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('name');
             $table->date('date');
             $table->string('url');
-            $table->integer('max');
-            $table->integer('min');
+            $table->unsignedTinyInteger('max');
+            $table->unsignedTinyInteger('min');
             $table->timestamps();
         });
     }
