+++ database/migrations/2023_03_28_072103_create_companies_table.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->string('ubication');
             $table->string('email');
             $table->string('phone');
-            $table->integer('priority');
+            $table->unsignedTinyInteger('priority');
 
             $table->timestamps();
         });
